281: move plan classes into plan/class/
